WIP make it build - this pretty much touches everything

+++ tools/libxc/xg_save_restore.h
@@ -24,7 +24,7 @@
 ** We process save/restore/migrate in batches of pages; the below
 ** determines how many pages we (at maximum) deal with in each batch.
 */
-#define MAX_PAGE_BATCH_SIZE 1024   /* up to 1024 pages (4MB) at a time */
+#define MAX_PAGE_BATCH_SIZE ((size_t)1024U)   /* up to 1024 pages (4MB) */
 
 /*
 ** We process the migration postcopy transition in batches of pfns to ensure
@@ -38,7 +38,7 @@
 ** The postcopy background copy uses a smaller batch size to ensure it can
 ** quickly respond to remote faults.
 */
-#define MAX_POSTCOPY_BATCH_SIZE 64
+#define MAX_POSTCOPY_BATCH_SIZE ((size_t)64U)
 
 /* When pinning page tables at the end of restore, we also use batching. */
 #define MAX_PIN_BATCH  1024
